requirements.txt
South==0.7.4
factory-boy==1.2.0
nose==1.2.1
django-nose==1.1
pinocchio==0.3.1
django-debug-toolbar==0.9.4
django-model-utils==1.1.0
django-braces==0.1.0

Django==1.5c1
